+++ src/main/java/org/betonquest/betonquest/quest/condition/time/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * {@link org.betonquest.betonquest.api.quest.condition.nullable.NullableCondition NullableCondition}
+ * Contains classes for time conditions.
+ */
+package org.betonquest.betonquest.quest.condition.time;