+++ src/main/java/org/betonquest/betonquest/quest/registry/CoreQuestTypes.java
@@ -43,7 +43,6 @@ import org.betonquest.betonquest.conditions.RideCondition;
 import org.betonquest.betonquest.conditions.ScoreboardCondition;
 import org.betonquest.betonquest.conditions.SneakCondition;
 import org.betonquest.betonquest.conditions.StageCondition;
-import org.betonquest.betonquest.conditions.TimeCondition;
 import org.betonquest.betonquest.objectives.ActionObjective;
 import org.betonquest.betonquest.objectives.ArrowShootObjective;
 import org.betonquest.betonquest.objectives.BlockObjective;
@@ -83,6 +82,7 @@ import org.betonquest.betonquest.quest.condition.realtime.DayOfWeekConditionFact
 import org.betonquest.betonquest.quest.condition.realtime.PartialDateConditionFactory;
 import org.betonquest.betonquest.quest.condition.tag.GlobalTagConditionFactory;
 import org.betonquest.betonquest.quest.condition.tag.TagConditionFactory;
+import org.betonquest.betonquest.quest.condition.time.TimeConditionFactory;
 import org.betonquest.betonquest.quest.condition.variable.VariableConditionFactory;
 import org.betonquest.betonquest.quest.condition.weather.WeatherConditionFactory;
 import org.betonquest.betonquest.quest.condition.world.WorldConditionFactory;
@@ -267,7 +267,7 @@ public class CoreQuestTypes {
         conditionTypes.register("stage", StageCondition.class);
         conditionTypes.register("tag", new TagConditionFactory(betonQuest));
         conditionTypes.registerCombined("testforblock", new BlockConditionFactory(data));
-        conditionTypes.register("time", TimeCondition.class);
+        conditionTypes.registerCombined("time", new TimeConditionFactory(data, betonQuest.getVariableProcessor()));
         conditionTypes.register("variable", new VariableConditionFactory(loggerFactory, data, betonQuest.getVariableProcessor()));
         conditionTypes.registerCombined("weather", new WeatherConditionFactory(data, betonQuest.getVariableProcessor()));
         conditionTypes.register("world", new WorldConditionFactory(loggerFactory, data, betonQuest.getVariableProcessor()));